remote queries: upload directly from the integration, drop the Go relay
Replace the Agent-owned HTTP upload relay with control-plane forwarding:
the Postgres Python integration now uploads bounded COPY chunks directly to
its-agent-intake over HTTP, eliminating the Go-side relay and the per-chunk
Python->C->Go copy.

ExecuteStream no longer intercepts data events; it passes the emit callback
straight through to the integration. validateRemoteQueryResultDelivery keeps
the mode/format/caps checks but drops the base-URL allowlist and redirect
rejection (the intake mints and owns the URL). The Agent now forwards
baseUrl and token to the integration in resultDelivery instead of sanitizing
them out; the integration reads the org API key and POC application key from
Agent config via datadog_agent.get_config and attaches them to its own HTTP
upload requests.

Delete remote_query_upload.go and its test (the Go relay, transport, retry,
finalize, and abort logic) and drop the pkg/util/http dependency. Update the
grpcserver proto-conversion test to assert the base URL is preserved
(forwarded) rather than rejected.

The intake data plane stays on /api/unstable/its-agent-intake/... +
ValidFullAPIUser (API key + application key + scoped upload token) for the
POC; API-key-only /api/intake + AAA approval remains a later production phase.

(cherry picked from commit 30e8ab5e4dc30004a7892532b5250c7180caaa4b)
